fix compile defintions and rename jit
