+++ src/coreclr/jit/alloc.cpp
@@ -330,7 +330,7 @@ void ArenaAllocator::dumpMaxMemStats(FILE* file)
 
 #ifdef JIT_STANDALONE_BUILD
 
-#if !defined(TARGET_LLVM) 
+#if !defined(TARGET_LLVM)
 void* __cdecl operator new(std::size_t size)
 {
     assert(!"Global new called; use HostAllocator if long-lived allocation was intended");
