+++ src/coreclr/jit/indirectcalltransformer.cpp
@@ -473,13 +473,13 @@ private:
 
     private:
         // Wasm stores function pointers as indices in a function table
-#if defined(TARGET_WASM32)
+#if defined(TARGET_LLVM_WASM32)
         const int FAT_POINTER_MASK = 0x80000000;
-#elif defined(TARGET_WASM64)
+#elif defined(TARGET_LLVM_WASM64)
         const int FAT_POINTER_MASK = 0x8000000000000000;
 #else
         const int FAT_POINTER_MASK = 0x2;
-#endif // TARGET_WASM32
+#endif // TARGET_LLVM_WASM32
 
         GenTree*  fptrAddress;
         var_types pointerType;
