+++ src/coreclr/jit/target.h
@@ -56,9 +56,9 @@ inline bool compUnixX86Abi()
 #define TARGET_READABLE_NAME "ARM"
 #elif defined(TARGET_ARM64)
 #define TARGET_READABLE_NAME "ARM64"
-#elif defined(TARGET_WASM64)
+#elif defined(TARGET_LLVM_WASM64)
 #define TARGET_READABLE_NAME "WASM64"
-#elif defined(TARGET_WASM32)
+#elif defined(TARGET_LLVM_WASM32)
 #define TARGET_READABLE_NAME "WASM32"
 #elif defined(TARGET_LOONGARCH64)
 #define TARGET_READABLE_NAME "LOONGARCH64"
@@ -339,7 +339,7 @@ public:
     }
 #endif
 
-#if !defined(TARGET_X86) && !defined(TARGET_WASM32) 
+#if !defined(TARGET_X86) && !defined(TARGET_LLVM_WASM32)
     explicit operator unsigned int() const
     {
         return (unsigned int)low;
@@ -604,7 +604,7 @@ static uint32_t BitScanForward(const regMaskTP& mask)
 #elif defined(TARGET_ARM64)
 #include "targetarm64.h"
 #elif defined(TARGET_LLVM)
-#include "targetwasm.h"
+#include "targetllvm.h"
 #elif defined(TARGET_LOONGARCH64)
 #include "targetloongarch64.h"
 #elif defined(TARGET_RISCV64)
@@ -632,7 +632,7 @@ static uint32_t BitScanForward(const regMaskTP& mask)
 
 #endif // TARGET_XARCH
 
-#if !defined(TARGET_WASM32) && !defined(TARGET_WASM64) // has no registers
+#if !defined(TARGET_LLVM_WASM32) && !defined(TARGET_LLVM_WASM64) // has no registers
 static_assert(REG_FIRST == 0);
 static_assert(REG_INT_FIRST < REG_INT_LAST);
 static_assert(REG_FP_FIRST  < REG_FP_LAST);
@@ -1147,12 +1147,12 @@ inline bool isFloatRegType(var_types type)
 
 // If the WINDOWS_AMD64_ABI is defined make sure that TARGET_AMD64 is also defined.
 #if defined(WINDOWS_AMD64_ABI)
-#if !defined(TARGET_AMD64) && !defined(TARGET_WASM32) && !defined(TARGET_WASM64)
+#if !defined(TARGET_AMD64) && !defined(TARGET_LLVM_WASM32) && !defined(TARGET_LLVM_WASM64)
 #error When WINDOWS_AMD64_ABI is defined you must define TARGET_AMD64 defined as well.
 #endif
 #endif
 
-#if !defined(TARGET_WASM32) && !defined(TARGET_WASM64)
+#if !defined(TARGET_LLVM_WASM32) && !defined(TARGET_LLVM_WASM64)
 // RBM_ALLINT is not known at compile time on TARGET_AMD64 since it's dependent on APX support.
 // Check should still be functional minus eGPR registers
 /*****************************************************************************/
